% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{closeEnough}
\alias{closeEnough}
\title{Check if two numbers are close enough per tolerance set by \code{testpack2_defaults$tol}}
\usage{
closeEnough(x, y)
}
\arguments{
\item{x, }{first value}

\item{y, }{second value}
}
\value{
TRUE or FALSE depending on \code{testpack2} package tolerance specification
}
\description{
Check if two numbers are close enough per tolerance set by \code{testpack2_defaults$tol}
}
\examples{
closeEnough(1.0001, 1.0002) ##TRUE per default tolerance of 1e-3
closeEnough(1.01, 1.02) ##FALSE per default tolerance of 1e-3
setDefaults("tol", 1e-1)
closeEnough(1.01, 1.02) ##TRUE per default tolerance of 1e-1

}
