% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{countOverlap}
\alias{countOverlap}
\title{Count overlap to see how well the reads were merged}
\usage{
countOverlap(merged_reads)
}
\arguments{
\item{merged_reads}{}
}
\value{

}
\description{
Count overlap to see how well the reads were merged
}
