% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cutadapt_functions.R
\name{cutadapt_tibble}
\alias{cutadapt_tibble}
\title{Prepare for primmer trimming with Cutaapt. Make new sub-directories and specify paths for the trimmed and untrimmed reads}
\usage{
cutadapt_tibble(fastq_data, metadata, intermediate_path)
}
\arguments{
\item{intermediate_path}{}
}
\value{

}
\description{
Prepare for primmer trimming with Cutaapt. Make new sub-directories and specify paths for the trimmed and untrimmed reads
}
