% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{dada2_readcounts_multi_sample}
\alias{dada2_readcounts_multi_sample}
\title{Final inventory of read counts after each step from input to removal of chimeras. This function deals with if you have more than one sample. TODO optimize for one sample}
\usage{
dada2_readcounts_multi_sample(asv_abund_table)
}
\value{

}
\description{
Final inventory of read counts after each step from input to removal of chimeras. This function deals with if you have more than one sample. TODO optimize for one sample
}
