% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_prepfunctions.R
\name{filter_and_trim}
\alias{filter_and_trim}
\title{Wrapper function for filterAndTrim function from DADA2 after primer removal}
\usage{
filter_and_trim(intermediate_path, cutadapt_data)
}
\arguments{
\item{cutadapt_data}{}
}
\value{

}
\description{
Wrapper function for filterAndTrim function from DADA2 after primer removal
}
