% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{format_abund_matrix}
\alias{format_abund_matrix}
\title{Format ASV abundance table}
\usage{
format_abund_matrix(asv_abund_table, seq_tax_asv)
}
\arguments{
\item{asv_abund_table}{}
}
\value{

}
\description{
Format ASV abundance table
}
