% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_databases.R
\name{format_database}
\alias{format_database}
\title{Create modified reference rps10 database for downstream analysis}
\usage{
format_database(raw_data_path, database_name)
}
\arguments{
\item{database_name}{}
}
\value{
A rps10 database that has modified headers and is output in the reference_databases folder.
}
\description{
Create modified reference rps10 database for downstream analysis
}
