% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{infer_asv_command}
\alias{infer_asv_command}
\title{Wrapper function to infer ASVs, for multiple loci}
\usage{
infer_asv_command(intermediate_path, cutadapt_data)
}
\arguments{
\item{denoised_data_path}{}
}
\value{

}
\description{
Wrapper function to infer ASVs, for multiple loci
}
