% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{infer_asvs}
\alias{infer_asvs}
\title{Core DADA2 function to learn errors and infer ASVs}
\usage{
infer_asvs(my_primer_pair_id, my_direction)
}
\arguments{
\item{my_direction}{}
}
\value{

}
\description{
Core DADA2 function to learn errors and infer ASVs
}
