% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{make_abund_table}
\alias{make_abund_table}
\title{Quality filtering to remove chimeras and short sequences}
\usage{
make_abund_table(raw_seqtab)
}
\arguments{
\item{seqtab}{}
}
\value{

}
\description{
Quality filtering to remove chimeras and short sequences
}
