% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{make_seqhist}
\alias{make_seqhist}
\title{Plots a histogram of read length counts of all sequences within the ASV table}
\usage{
make_seqhist(abund_table, abund_table_file)
}
\arguments{
\item{abund_table_file}{}
}
\value{

}
\description{
Plots a histogram of read length counts of all sequences within the ASV table
}
