% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{merge_reads_command}
\alias{merge_reads_command}
\title{Merge forward and reverse reads}
\usage{
merge_reads_command(intermediate_path)
}
\arguments{
\item{merged_read_data_path}{}
}
\value{

}
\description{
Merge forward and reverse reads
}
