% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_prepfunctions.R
\name{plot_qc}
\alias{plot_qc}
\title{Wrapper function for plotQualityProfile function}
\usage{
plot_qc(cutadapt_data, intermediate_path)
}
\value{

}
\description{
Wrapper function for plotQualityProfile function
}
