% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotbymonth}
\alias{plotbymonth}
\title{Plot ufo sightings by month}
\usage{
plotbymonth(state = "NV")
}
\arguments{
\item{state}{state to use
Default is "NV"}
}
\value{
returns a plot.  See \link{ufo}
}
\description{
Plot ufo sightings by month
}
