% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primer_hits.R
\name{post_primer_hit_plot}
\alias{post_primer_hit_plot}
\title{Make a barplot of primers identified on reads after trim steps}
\usage{
post_primer_hit_plot(primer_hits, fastq_data, intermediate_path, plot_name)
}
\arguments{
\item{metadata}{}
}
\value{

}
\description{
Make a barplot of primers identified on reads after trim steps
}
