% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_prepfunctions.R
\name{post_trim_qc}
\alias{post_trim_qc}
\title{Wrapper script for plotQualityProfile after trim steps and primer removal.}
\usage{
post_trim_qc(cutadapt_data, intermediate_path)
}
\arguments{
\item{intermediate_path}{}
}
\value{

}
\description{
Wrapper script for plotQualityProfile after trim steps and primer removal.
}
