% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primer_hits.R
\name{pre_primer_hit_data}
\alias{pre_primer_hit_data}
\title{Get primer counts fo reach sample before primer removal and trimming steps}
\usage{
pre_primer_hit_data(primer_data, fastq_data, intermediate_path)
}
\arguments{
\item{intermediate_path}{}
}
\value{

}
\description{
Get primer counts fo reach sample before primer removal and trimming steps
}
