% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_trim_functions.R
\name{prepare_fastq}
\alias{prepare_fastq}
\title{A function for calling read_fastq, matching_order_primer_check, and remove_ns functions. This will process and edit the FASTQ and make them ready for the trimming of primers with Cutadapt.}
\usage{
prepare_fastq(raw_path, intermediate_path)
}
\arguments{
\item{metadata}{}
}
\value{

}
\description{
A function for calling read_fastq, matching_order_primer_check, and remove_ns functions. This will process and edit the FASTQ and make them ready for the trimming of primers with Cutadapt.
}
