% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_primers.R
\name{prepare_primers}
\alias{prepare_primers}
\title{Take in user primers and creates the complement, reverse, reverse complement of primers in one tibble}
\usage{
prepare_primers(primer_path)
}
\arguments{
\item{primer_path}{a path to the csv file that holds the primer info for this project}
}
\value{
A tibble that contains the forward and reverse primers with all complements of primers
}
\description{
Take in user primers and creates the complement, reverse, reverse complement of primers in one tibble
}
