% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_prepfunctions.R
\name{remove_ns}
\alias{remove_ns}
\title{Wrapper function for core DADA2 filter and trim function for first filtering step}
\usage{
remove_ns(fastq_data, intermediate_path)
}
\value{

}
\description{
Wrapper function for core DADA2 filter and trim function for first filtering step
}
