% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{resetDefaults}
\alias{resetDefaults}
\title{Function to set defaults}
\usage{
resetDefaults(tol = 0.001, max_iter = 25)
}
\arguments{
\item{tol}{the tolerance, default 1e-3}

\item{max_iter}{the maximum number of iterations, default 25}
}
\value{
a list of default values invisibly
}
\description{
Function to set defaults
}
\examples{
(resetDefaults())
(resetDefaults(tol = 1e-5))
}
