% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2core_functions.R
\name{separate_abund_table}
\alias{separate_abund_table}
\title{Still needs improvement. Could solve problem by merging fungal and oomycete databases for assign taxonomy steps. In order to assign taxonomy for separate loci, need to separate abundance tables first}
\usage{
separate_abund_table(
  abund_asv_its,
  abund_asv_rps10,
  intermediate_path,
  separate_abund_path
)
}
\arguments{
\item{separate_abund_path}{}
}
\value{

}
\description{
Still needs improvement. Could solve problem by merging fungal and oomycete databases for assign taxonomy steps. In order to assign taxonomy for separate loci, need to separate abundance tables first
}
