% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setDefaults}
\alias{setDefaults}
\title{Set the default param name to a value}
\usage{
setDefaults(name, value)
}
\arguments{
\item{name}{the parameter name}

\item{value}{the value to assign}
}
\value{
the new list of defaults invisibly
}
\description{
Set the default param name to a value
}
\examples{
(setDefaults("tol", 1e-5))
(setDefaults("max_iter", 100))
}
