\name{TestPack2-package}
\alias{TestPack2-package}
\alias{TestPack2}
\docType{package}
\title{
\packageTitle{TestPack2}
}
\description{
\packageDescription{TestPack2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TestPack2}
\packageIndices{TestPack2}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{TestPack2}

Maintainer: \packageMaintainer{TestPack2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
