% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpack2.R
\docType{package}
\name{testpack2}
\alias{testpack2}
\title{testpack2: Learnin' about UFOs}
\description{
Some description here
}
\section{Data import functions}{

\code{\link{ufo_data}}
}

\section{Plotting functions}{

\code{\link{plotbymonth}}
}

\section{Other section can go here, e.g. thoughts on future dev}{

\itemize{
  \item first item
  \item second item
}
}

